feat: creation of Three Enums.
First enum 'CourseLevel' created with the model browser with its three values.

Second enum 'Level' which is declared within the main program with its three values.

Finally, third enum created as empoty using the model browser, but this one is referenced to the 'Level' enum created within the main program, so the empty one takes the values from the other one.